24.09.06 수업내용(half)

+++ 01_variable/src/edu/kh/variable/practice/CastingPractice1.java
@@ -18,7 +18,7 @@ public class CastingPractice1 {
 		
 		System.out.println( iNum1/(double)iNum2 ); // 2.5
 		
-		System.out.println( dNum ); // 2.5
+		System.out.println( (int)dNum ); // 2.5
 		
 		System.out.println( (int)fNum ); // 3
